feat: connect testbed and orchestrator to store checker results
Move verification update function to trs2wes module and include
in orchestrator run monitoring to update queue config details.
