+++ synorchestrator/testbed.py
@@ -81,7 +81,8 @@ def check_workflow(queue_id, wes_id):
               wf_id=checker_id,
               version_id=wf_config['version_id'],
               wes_default=wf_config['wes_default'],
-              wes_opts=wf_config['wes_opts'])
+              wes_opts=wf_config['wes_opts'],
+              target_queue=queue_id)
 
     checker_job = trs_instance.get_workflow_tests(id=checker_id,
                                                   version_id=wf_config['version_id'],
@@ -104,21 +105,3 @@ def check_all(workflow_wes_map):
     return submission_logs
 
 
-def store_verification(queue_id, wes_id):
-    """
-    Record checker status for selected workflow and environment.
-    """
-    wf_config = queue_config()[queue_id]
-    wf_config.setdefault('wes_verified', []).append(wes_id)
-    set_yaml('queues', queue_id, wf_config)
-
-
-# def post_verification(self, id, version_id, type, relative_path, requests):
-#     """
-#     Annotate test JSON with information on whether it ran successfully on particular platforms plus metadata
-#     """
-#     id = _format_workflow_id(id)
-#     endpoint ='extended/{}/versions/{}/{}/tests/{}'.format(
-#         id, version_id, type, relative_path
-#     )
-#     return _post_to_endpoint(self, endpoint, requests)
